"Removed AdminDashboard component and refactored AdminLayout; added createEstimate and updateEstimateStatus mutations to convex/estimates.ts"
